功能完成 添加了dao层和service大量的JPA函数相关和seedBag买入逻辑和删除逻辑 完善controller层代码逻辑，主要负责数据预处理和转发 添加数据库表主外键约束和非主键唯一约束 完善页面数据交互 还需要修改按钮图片

+++ src/cn/jxufe/controller/ShopController.java
@@ -1,23 +1,15 @@
 package cn.jxufe.controller;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import cn.jxufe.bean.EasyUIData;
-import cn.jxufe.bean.EasyUIDataPageRequest;
 import cn.jxufe.bean.Message;
 import cn.jxufe.entity.SeedBag;
 import cn.jxufe.entity.User;
@@ -48,16 +40,38 @@ public class ShopController {
 			if(user!=null)id=user.getId();
 			return seedBagService.findByUId(id);
 	}
-	@RequestMapping(value = "save", produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "save", produces = MediaType.APPLICATION_JSON_VALUE,consumes=MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public Message save(SeedBag seedBag, Model model) {
-		return seedBagService.save(seedBag);
+	public Message save(@RequestBody SeedBag seedBag, HttpSession session,Model model) {
+		User user=(User)session.getAttribute("user");
+		if(user!=null) {
+			seedBag.setuId(user.getId());
+			return seedBagService.save(seedBag);
+		}
+		else {
+			Message message=new Message();
+			message.setCode(-100);
+			message.setMsg("未登录");
+			return message;
+		}
+		
 	}
 
-	@RequestMapping(value = "delete", produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "delete", produces = MediaType.APPLICATION_JSON_VALUE,consumes=MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public Message delete(SeedBag seedBag, Model model) {
-		return seedBagService.delete(seedBag);
+	public Message delete(@RequestBody SeedBag seedBag, HttpSession session,Model model) {
+		User user=(User)session.getAttribute("user");
+		if(user!=null) {
+			seedBag.setuId(user.getId());
+			return seedBagService.delete(seedBag);
+		}
+		else {
+			Message message=new Message();
+			message.setCode(-100);
+			message.setMsg("未登录");
+			return message;
+		}
+		
 	}
 
 }
